Fix deletions with multiple leading dashes
Deleted lines may start with the leading dash "-" but then the line
itself may contain an arbitrary number of dashes. This change parses
any number of leading dashes as a deletion except for the triple
leading dash which is considered to be an old file marker.

These cases should have been valid parse examples before but were
marked as errors because of the contention between the deletion
and old_file rules and error recovery.

Connects #1